+++ smartsim/_core/mli/infrastructure/control/workermanager.py
@@ -24,24 +24,34 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-import multiprocessing as mp
+import sys
+
+# isort: off
+import dragon
+from dragon import fli
+
+# isort: on
+
+import time
 import typing as t
 
 import numpy as np
 
-from smartsim._core.entrypoints.service import Service
-from smartsim._core.mli.comm.channel.channel import CommChannelBase
-from smartsim._core.mli.comm.channel.dragonchannel import DragonCommChannel
-from smartsim._core.mli.infrastructure.environmentloader import EnvironmentConfigLoader
-from smartsim._core.mli.infrastructure.storage.featurestore import FeatureStore
-from smartsim._core.mli.infrastructure.worker.worker import (
+from .....error import SmartSimError
+from .....log import get_logger
+from ....entrypoints.service import Service
+from ...comm.channel.channel import CommChannelBase
+from ...comm.channel.dragonchannel import DragonCommChannel
+from ...infrastructure.environmentloader import EnvironmentConfigLoader
+from ...infrastructure.storage.featurestore import FeatureStore
+from ...infrastructure.worker.worker import (
     InferenceReply,
     InferenceRequest,
+    LoadModelResult,
     MachineLearningWorkerBase,
 )
-from smartsim._core.mli.message_handler import MessageHandler
-from smartsim._core.mli.mli_schemas.response.response_capnp import Response
-from smartsim.log import get_logger
+from ...message_handler import MessageHandler
+from ...mli_schemas.response.response_capnp import Response
 
 if t.TYPE_CHECKING:
     from dragon.fli import FLInterface
@@ -53,7 +63,9 @@ logger = get_logger(__name__)
 
 
 def deserialize_message(
-    data_blob: bytes, channel_type: t.Type[CommChannelBase]
+    data_blob: bytes,
+    channel_type: t.Type[CommChannelBase],
+    device: t.Literal["cpu", "gpu"],
 ) -> InferenceRequest:
     """Deserialize a message from a byte stream into an InferenceRequest
     :param data_blob: The byte stream to deserialize"""
@@ -87,12 +99,6 @@ def deserialize_message(
         None  # these will really be tensors already
     )
 
-    # # client example
-    # msg = Message()
-    # t = torch.Tensor()
-    # msg.inputs = [custom_byte_converter(t)]
-    # mli_client.request_inference(msg)
-    # # end client
     input_meta: t.List[t.Any] = []
 
     if request.input.which() == "keys":
@@ -170,6 +176,7 @@ class WorkerManager(Service):
         as_service: bool = False,
         cooldown: int = 0,
         comm_channel_type: t.Type[CommChannelBase] = DragonCommChannel,
+        device: t.Literal["cpu", "gpu"] = "cpu",
     ) -> None:
         """Initialize the WorkerManager
         :param config_loader: Environment config loader that loads the task queue and
@@ -182,8 +189,7 @@ class WorkerManager(Service):
         """
         super().__init__(as_service, cooldown)
 
-        """a collection of workers the manager is controlling"""
-        self._task_queue: t.Optional["FLInterface"] = config_loader.get_queue()
+        self._task_queue: t.Optional[CommChannelBase] = config_loader.get_queue()
         """the queue the manager monitors for new tasks"""
         self._feature_store: t.Optional[FeatureStore] = (
             config_loader.get_feature_store()
@@ -193,6 +199,10 @@ class WorkerManager(Service):
         """The ML Worker implementation"""
         self._comm_channel_type = comm_channel_type
         """The type of communication channel to construct for callbacks"""
+        self._device = device
+        """Device on which workers need to run"""
+        self._cached_models: dict[str, t.Any] = {}
+        """Dictionary of previously loaded models"""
 
     def _validate_request(self, request: InferenceRequest) -> bool:
         """Ensure the request can be processed.
@@ -234,24 +244,68 @@ class WorkerManager(Service):
             logger.warning("No queue to check for tasks")
             return
 
+        timings = []  # timing
         # perform default deserialization of the message envelope
-        request_bytes: bytes = self._task_queue.get()
+        request_bytes: bytes = self._task_queue.recv()
 
-        request = deserialize_message(request_bytes, self._comm_channel_type)
+        interm = time.perf_counter()  # timing
+        request = deserialize_message(
+            request_bytes, self._comm_channel_type, self._device
+        )
         if not self._validate_request(request):
             return
 
-        # # let the worker perform additional custom deserialization
-        # request = self._worker.deserialize(request_bytes)
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
 
-        fetch_model_result = self._worker.fetch_model(request, self._feature_store)
-        model_result = self._worker.load_model(request, fetch_model_result)
+        if not request.raw_model:
+            if request.model_key is None:
+                # A valid request should never get here.
+                raise ValueError("Could not read model key")
+            if request.model_key in self._cached_models:
+                timings.append(time.perf_counter() - interm)  # timing
+                interm = time.perf_counter()  # timing
+                model_result = LoadModelResult(self._cached_models[request.model_key])
+
+            else:
+                fetch_model_result = None
+                while True:
+                    try:
+                        interm = time.perf_counter()  # timing
+                        fetch_model_result = self._worker.fetch_model(
+                            request, self._feature_store
+                        )
+                    except KeyError:
+                        time.sleep(0.1)
+                    else:
+                        break
+
+                if fetch_model_result is None:
+                    raise SmartSimError("Could not retrieve model from feature store")
+                timings.append(time.perf_counter() - interm)  # timing
+                interm = time.perf_counter()  # timing
+                model_result = self._worker.load_model(
+                    request, fetch_model_result, self._device
+                )
+                self._cached_models[request.model_key] = model_result.model
+        else:
+            fetch_model_result = self._worker.fetch_model(request, None)
+            model_result = self._worker.load_model(
+                request, fetch_result=fetch_model_result, device=self._device
+            )
+
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
         fetch_input_result = self._worker.fetch_inputs(request, self._feature_store)
-        transformed_input = self._worker.transform_input(request, fetch_input_result)
 
-        # batch: t.Collection[_Datum] = transform_result.transformed_input
-        # if self._batch_size:
-        #     batch = self._worker.batch_requests(transform_result, self._batch_size)
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
+        transformed_input = self._worker.transform_input(
+            request, fetch_input_result, self._device
+        )
+
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
 
         reply = InferenceReply()
 
@@ -260,8 +314,14 @@ class WorkerManager(Service):
                 request, model_result, transformed_input
             )
 
-            transformed_output = self._worker.transform_output(request, execute_result)
+            timings.append(time.perf_counter() - interm)  # timing
+            interm = time.perf_counter()  # timing
+            transformed_output = self._worker.transform_output(
+                request, execute_result, self._device
+            )
 
+            timings.append(time.perf_counter() - interm)  # timing
+            interm = time.perf_counter()  # timing
             if request.output_keys:
                 reply.output_keys = self._worker.place_output(
                     request, transformed_output, self._feature_store
@@ -272,6 +332,9 @@ class WorkerManager(Service):
             logger.exception("Error executing worker")
             reply.failed = True
 
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
+
         if reply.failed:
             response = build_failure_reply("fail", "failure-occurred")
         else:
@@ -280,10 +343,21 @@ class WorkerManager(Service):
 
             response = build_reply(reply)
 
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
+
         # serialized = self._worker.serialize_reply(request, transformed_output)
         serialized_resp = MessageHandler.serialize_response(response)  # type: ignore
+
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
         if request.callback:
             request.callback.send(serialized_resp)
+
+        timings.append(time.perf_counter() - interm)  # timing
+        interm = time.perf_counter()  # timing
+
+        print(" ".join(str(time) for time in timings))  # timing
 
     def _can_shutdown(self) -> bool:
         """Return true when the criteria to shut down the service are met."""